TIP-842: Move Group model to enrichment sub context
