+++ src/Pim/Bundle/CatalogBundle/Resources/config/doctrine/Group.orm.yml
@@ -15,7 +15,7 @@ Pim\Bundle\CatalogBundle\Entity\Group:
             unique: true
     manyToOne:
         type:
-            targetEntity: Pim\Component\Catalog\Model\GroupTypeInterface
+            targetEntity: Akeneo\Pim\Enrichment\Component\Product\Model\GroupTypeInterface
             inversedBy: groups
             joinColumns:
                 type_id:
