+++ src/Pim/Bundle/CatalogBundle/Resources/config/entities.yml
@@ -12,7 +12,7 @@ parameters:
     pim_catalog.entity.category_translation.class:         Pim\Bundle\CatalogBundle\Entity\CategoryTranslation
     pim_catalog.entity.group.class:                        Pim\Bundle\CatalogBundle\Entity\Group
     pim_catalog.entity.group_translation.class:            Pim\Bundle\CatalogBundle\Entity\GroupTranslation
-    pim_catalog.entity.group_type.class:                   Pim\Bundle\CatalogBundle\Entity\GroupType
+    pim_catalog.entity.group_type.class:                   Akeneo\Pim\Enrichment\Component\Product\Model\GroupType
     pim_catalog.entity.group_type_translation.class:       Pim\Bundle\CatalogBundle\Entity\GroupTypeTranslation
     pim_catalog.entity.attribute.class:                    Akeneo\Pim\Structure\Component\Model\Attribute
     pim_catalog.entity.attribute_translation.class:        Akeneo\Pim\Structure\Component\Model\AttributeTranslation
