+++ src/Pim/Bundle/CatalogBundle/Resources/config/models.yml
@@ -7,7 +7,7 @@ parameters:
     pim_catalog.model.association_type.interface:       Pim\Component\Catalog\Model\AssociationTypeInterface
     pim_catalog.model.category.interface:               Pim\Component\Catalog\Model\CategoryInterface
     pim_catalog.model.group.interface:                  Pim\Component\Catalog\Model\GroupInterface
-    pim_catalog.model.group_type.interface:             Pim\Component\Catalog\Model\GroupTypeInterface
+    pim_catalog.model.group_type.interface:             Akeneo\Pim\Enrichment\Component\Product\Model\GroupTypeInterface
     pim_catalog.model.attribute.interface:              Akeneo\Pim\Structure\Component\Model\AttributeInterface
     pim_catalog.model.attribute_group.interface:        Akeneo\Pim\Structure\Component\Model\AttributeGroupInterface
     pim_catalog.model.attribute_option.interface:       Akeneo\Pim\Structure\Component\Model\AttributeOptionInterface
