+++ src/Pim/Bundle/EnrichBundle/spec/Form/Type/GroupTypeTypeSpec.php
@@ -3,7 +3,7 @@
 namespace spec\Pim\Bundle\EnrichBundle\Form\Type;
 
 use PhpSpec\ObjectBehavior;
-use Pim\Bundle\CatalogBundle\Entity\GroupType;
+use Akeneo\Pim\Enrichment\Component\Product\Model\GroupType;
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\OptionsResolver\OptionsResolver;
