+++ src/Pim/Component/Catalog/spec/Normalizer/Standard/GroupTypeNormalizerSpec.php
@@ -3,7 +3,7 @@
 namespace spec\Pim\Component\Catalog\Normalizer\Standard;
 
 use PhpSpec\ObjectBehavior;
-use Pim\Component\Catalog\Model\GroupTypeInterface;
+use Akeneo\Pim\Enrichment\Component\Product\Model\GroupTypeInterface;
 use Pim\Component\Catalog\Normalizer\Standard\TranslationNormalizer;
 
 class GroupTypeNormalizerSpec extends ObjectBehavior
